LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-base-expected.txt


FAIL flip-start affects base values assert_equals: expected "300px" but got "150px"
